pyproject.toml
[project]
name = "discord2notion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.7.0",
    "notion-client>=3.0.0",
    "python-dotenv>=1.2.2",
]
